it will print log when an exception occurs to an HTTP request。Output information includes IP、error_msg、url、request_param ...
