+++ pom.xml
@@ -316,11 +316,6 @@
 					</exclusion>
 				</exclusions>
 			</dependency>
-			<dependency>
-				<groupId>com.alibaba</groupId>
-				<artifactId>fastjson</artifactId>
-				<version>1.2.51</version>
-			</dependency>
 		</dependencies>
 	</dependencyManagement>
